+++ docs/releasenotes/releasenotes.rst
@@ -13,6 +13,7 @@ Enhancements
 
 * Calendar: New block parameter ``emptytext`` to define the text to show where there are no calendar appointments. Only works for the new calendar block. See :ref:`newcalendar`
 * Custom graph: aggregate parameter can be an array to specify different aggregation methods per data element. See :ref:`groupBy`
+* Support for device (sub)type Managed Counter
 
 Fixes
 ~~~~~~